pyproject.toml
[project]
name = "NetworkSecurity"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "numpy>=2.1.2",
    "pymongo>=3.11.0",
    "pymongo[srv]==3.11",
    "mlflow>=2.17.1",
    "scikit-learn>=1.5.2",
    "pyaml>=24.9.0",
    "dagshub>=0.3.40",
    "fastapi>=0.115.4",
    "python-multipart>=0.0.16",
    "certifi>=2024.8.30",
    "uvicorn>=0.32.0",
    "setuptools>=75.2.0",
    "from-root>=1.3.0",
    "pip>=24.3.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.build]
package-dir = "src/networksecurity"

# editables backend doesn't work well with namespace packages
#run-setuptools = true
editable-backend = "path"


[build-system]
requires = ["pdm-backend", "pdm-build-locked"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
